+++ src/canvas-effects/canvas-effect.js
@@ -630,10 +630,7 @@ export default class CanvasEffect extends PIXI.Container {
 			game.settings.get("sequencer", "effectsEnabled") &&
 			game.user.viewedScene === this.data.sceneId;
 
-		if (
-			foundry.utils.isNewerVersion(game.version, "10.289") &&
-			game.settings.get("core", "photosensitiveMode")
-		) {
+		if (game.settings.get("core", "photosensitiveMode")) {
 			playVisible = false;
 			lib.throttled_custom_warning(
 				this.data.moduleName,
@@ -1939,10 +1936,10 @@ export default class CanvasEffect extends PIXI.Container {
 		);
 		if (CONSTANTS.INTEGRATIONS.ISOMETRIC.ACTIVE && isIsometric) {
 			const sourceSort = this.source
-				? this.sourceMesh.sort + (this.data.isometric?.overlay ? 1 : -1)
+				? (this.sourceMesh?.sort ?? 0) + (this.data.isometric?.overlay ? 1 : -1)
 				: 0;
 			const targetSort = this.target
-				? this.targetMesh.sort + (this.data.isometric?.overlay ? 1 : -1)
+				? (this.targetMesh?.sort ?? 0) + (this.data.isometric?.overlay ? 1 : -1)
 				: 0;
 			this.sort = Math.max(sourceSort, targetSort);
 		} else {