+++ src/modules/sequencer.js
@@ -142,23 +142,11 @@ export default class Sequence {
       );
     }
 
-    if (foundry.utils.isNewerVersion(game.version, "11")) {
-      args = args.length ? args?.[0] : {};
-      if (typeof args !== "object") {
-        throw lib.custom_error(
-          this.moduleName,
-          `macro - Secondary argument must be an object`
-        );
-      }
-    } else if (
-      args &&
-      args.length &&
-      !game.modules.get("advanced-macros")?.active
-    ) {
-      lib.custom_warning(
+    args = args.length ? args?.[0] : {};
+    if (typeof args !== "object") {
+      throw lib.custom_error(
         this.moduleName,
-        `macro - Supplying macros with arguments require the advanced-macros module to be active`,
-        true
+        `macro - Secondary argument must be an object`
       );
     }
 
@@ -183,26 +171,7 @@ export default class Sequence {
             macro.ownership.default = CONST.DOCUMENT_OWNERSHIP_LEVELS.OWNER;
           }
 
-          if (foundry.utils.isNewerVersion(game.version, "11")) {
-            await macro.execute(args);
-          } else {
-            const version = game.modules.get("advanced-macros")?.version;
-            const bugAdvancedMacros =
-              game.modules.get("advanced-macros")?.active &&
-              foundry.utils.isNewerVersion(
-                version.startsWith("v") ? version.slice(1) : version,
-                "1.18.2"
-              ) &&
-              !foundry.utils.isNewerVersion(
-                version.startsWith("v") ? version.slice(1) : version,
-                "1.19.1"
-              );
-            if (bugAdvancedMacros) {
-              await macro.execute([...args]);
-            } else {
-              await macro.execute(...args);
-            }
-          }
+          await macro.execute(args);
         },
         true
       )